Corrected docs and tax helpers moved to module
